move screenshot and video capture functionality to own module
